+++ logs/CHANGELOG.md
@@ -1,4 +1,7 @@
 # 改动日志
+## 2026-05-07 21:06 | auto: 05-07 21:06 | src/screens/LessonScreen.tsx 
+- src/screens/LessonScreen.tsx
+
 ## 2026-05-07 21:00 | auto: 05-07 21:00 | src/screens/LessonScreen.tsx src/utils/animationHooks.ts 
 - src/screens/LessonScreen.tsx
 - src/utils/animationHooks.ts